requirements.txt
asgiref==3.3.1
astroid==2.5
asttokens==2.0.5
backcall==0.2.0
black==22.1.0
certifi==2020.12.5
chardet==4.0.0
click==8.0.4
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
Django==3.1.7
django-cors-headers==3.7.0
django-rest-swagger==2.2.0
djangorestframework==3.12.2
djangorestframework-camel-case==1.3.0
drf-yasg==1.20.0
executing==0.8.2
furl==2.1.0
idna==2.10
inflection==0.5.1
ipython==7.16.3
isort==5.7.0
itypes==1.2.0
jedi==0.17.2
Jinja2==2.11.3
lazy-object-proxy==1.5.2
MarkupSafe==1.1.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mypy-extensions==0.4.3
openapi-codec==1.3.2
orderedmultidict==1.0.1
packaging==20.9
parso==0.7.1
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
platformdirs==2.4.0
prompt-toolkit==3.0.28
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.11.2
pylint==2.7.0
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
simplejson==3.17.6
six==1.15.0
sqlparse==0.4.1
stack-data==0.2.0
toml==0.10.2
tomli==1.2.3
traitlets==4.3.3
typed-ast==1.4.2
typing_extensions==4.1.1
uritemplate==3.0.1
urllib3==1.26.3
wcwidth==0.2.5
wrapt==1.12.1
